Moving run_scenario to scenario.py
Fixing scenario run

Fix for import section
