+++ __init__.py
@@ -1,6 +1,8 @@
+"""
 from . import constants
 from . import utils
 from . import multi_partner_learning
 from . import contributivity
 from . import scenario
 from . import dataset
+"""
